LayoutTests/fast/events/message-port-no-wrapper-expected.txt

CONSOLE MESSAGE: line 26: ReferenceError: Can't find variable: MessageChannel
FAIL: Timed out waiting for notifyDone to be called
Test that delivering a message to a port that was never referenced from JS doesn't cause a crash.
